refactor: Extract ortholog tools (DIOPT) into separate module
- Created src/tools/ortholog_tools.py with 2 DIOPT functions:
  - get_diopt_orthologs() - Find orthologs across model organisms
  - get_diopt_alignment() - Get protein sequence alignments
- Updated src/tools/__init__.py to export ortholog_tools
- Updated server.py to import and register ortholog tools
- Removed ortholog tool implementations from server.py (~67 lines)
- server.py now ~220 lines (down from ~287 lines)

Part of PR #5 in refactoring series (see REFACTORING_PLAN.md)
